requirements.txt
RPi.GPIO
PyYAML
adafruit-circuitpython-ads1x15